Frontend/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
openai>=1.57.0
httpx>=0.27.0



